backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pymysql==1.0.2
pyjwt==2.1.0
python-dotenv==0.19.0
flask-cors==3.0.10
flask-uploads==0.2.1
python-magic==0.4.24
flask-migrate==3.1.0
pytest==7.0.1